requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.0
imageio==2.31.6
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.1
networkx==3.2.1
numpy==1.26.1
packaging==23.2
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
termcolor==2.3.0
tifffile==2023.9.26
